Update to Vert.x 4 EDGCOMMON-30 (#38)
Update to Vert.x 4.0.0 EDGCOMMON-30

The OkapiClient in this module (one of dozens) has changed its API.
OkapiClient.{post,put,get,delete} has a response handler now of type
Handler<HttpResponse<Buffer>> it used to be: Handler<HttpClientResponse>.

The OkapiClient really doesn't give much over WebClient. And
we should consider just dropping it.

Use log4j2 formatting; change POST to use info

+++ src/main/java/org/folio/edge/core/Constants.java
@@ -32,7 +32,7 @@ public class Constants {
   public static final String DEFAULT_SECURE_STORE_TYPE = "ephemeral";
   public static final String DEFAULT_PORT = "8081";
   public static final String DEFAULT_LOG_LEVEL = "INFO";
-  public static final long DEFAULT_REQUEST_TIMEOUT_MS = 30 * 1000L; // ms
+  public static final int DEFAULT_REQUEST_TIMEOUT_MS = 30 * 1000; // ms
   public static final long DEFAULT_TOKEN_CACHE_TTL_MS = 60 * 60 * 1000L;
   public static final long DEFAULT_NULL_TOKEN_CACHE_TTL_MS = 30 * 1000L;
   public static final int DEFAULT_TOKEN_CACHE_CAPACITY = 100;
